requirements.txt
streamlit
pandas
sqlalchemy
azure-identity
pyodbc
urllib3
azure-storage-file-datalake
deltalake
numpy